Implemented cache! to cache expensive data
This is going to be optimized later to use a fetch multi algorithm and
strings instead of marshalling.

We aren't doing any marshalling here but cache implementations always do
that, like active support cache stores.

By using strings instead of objects we can achieve higher throughput.

The expected interface of the cache store is this:

    cache_store.fetch(key, options, &block)

As you can guess, this is the interface implemented by active support,
so it should be plug n play for rails.

For other cache interfaces, you can implement a driver that have this
interface and maps to your desired interface.