Use fetch_multi instead of fetch
This will improve the performance since we are going to hit the backend
just once to read all keys, as long the cache adapter implements fetch
multi support like dalli.

For example:

    builder.cache! :x do |cache|
      cache.x true
    end

    builder.cache! :y do |cache|
      cache.y true
    end

    builder.cache! :z do |cache|
      cache.z true
    end

This example was hitting the memcached 6 times on cache miss:

  1. read x
  2. write x
  3. read y
  4. write y
  5. read z
  6. write z

And 3 times on cache hit:

  1. read x
  2. read y
  3. read z

After this change, 4 times on cache miss:

  1. read multi x,y,z
  2. write x
  3. write y
  4. write z

And 1 time on cache hit:

  1. read multi x,y,z

Note that in the case of different options, one read multi will be made
per each options, i.e.:

    builder.cache! :x do |cache|
      cache.x true
    end

    builder.cache! :y do |cache|
      cache.y true
    end

    builder.cache! :z, expires_in: 10.minutes do |cache|
      cache.z true
    end

    builder.cache! :w, expires_in: 10.minutes do |cache|
      cache.w true
    end

In the case of cache miss:

  1. read multi x,y
  2. write x
  3. write y
  4. read multi z,w
  5. write z
  5. write w

In the case of cache hit:

  1. read multi x,y
  2. read multi z,w

That's because Rails.cache.fetch_multi signature is limited to use the
same options for all given keys.

And for last, nested cache calls are allowed and will follow recursively
to accomplish the same behavior, i.e.:

  builder.cache! :x do |cache_x|
    cache_x.x true

    cache_x.cache! :y do |cache_y|
      cache_y.y true
    end

    cache_x.cache! :z do |cache_z|
      cache_z.z true
    end
  end

  builder.cache! :w do |cache_w|
    cache_w.w true
  end

In the case of cache miss:

  1. read multi x,w
  2. read multi y,z
  3. write y
  4. write z
  5. write x
  6. write w

In the case of cache hit:

  1. read multi x,w

The same rule of options will be applied, if you have different options,
one hit per options.

Pretty much the same of https://github.com/rails/jbuilder/pull/421.